Create FileImport-Exclusions_withPlantLocations
This code was made to upload GPR data and upload the plant locations, using them to exclude the "noisy" data outside of a certain radius. The plant locations have to be put in manually. Plant locations should be in the form ID, X, Y, Z. GPR data is in the form X, Y, Z, and Radar Return Values.